Refact relations helpers
