+++ src/ui/units/dash/containers/Dialogs/DialogRelations/test/helpers.test.ts
@@ -1,7 +1,8 @@
 import {Config} from '@gravity-ui/dashkit/build/esm/shared/types/config';
-import {RelationType} from 'ui/units/dash/containers/Dialogs/DialogRelations/types';
 
-import {RELATION_TYPES, addAlias, getNormalizedAliases, getUpdatedRelations} from '../helpers';
+import {RELATION_TYPES} from '../constants';
+import {addAlias, getNormalizedAliases, getUpdatedRelations} from '../helpers';
+import {RelationType} from '../types';
 
 jest.mock('ui', () => {
     return {
